requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Werkzeug==2.3.7
Jinja2==3.1.2
MarkupSafe==2.1.3
itsdangerous==2.1.2
click==8.1.7
blinker==1.6.3
openai>=1.3.0
python-dotenv==1.0.0
gunicorn==21.2.0
pandas>=2.0.0
openpyxl>=3.1.0
PyPDF2>=3.0.0
python-docx>=1.0.0